expand to other modules

+++ model/luepcr/operations.py
@@ -158,7 +158,7 @@ def boolean(arg):
         return lfr.valid(lfr.create_array(array_shape, partition_shape, np.dtype(np.float32), fill_value = 1.0))
         # ~ return lfr.create_array(array_shape, partition_shape, np.dtype(np.bool_), fill_value = 1.0)
     elif arg == 0.0:
-        return lfr.valid(lfr.create_array(array_shape, partition_shape, np.dtype(np.float32), fill_value = 0.0)) * lfr.create_array(array_shape, partition_shape, np.dtype(np.float32), fill_value = 0.0)
+        return lfr.valid(lfr.create_array(array_shape, partition_shape, np.dtype(np.float32), fill_value = 0.0)) * lfr.create_array(array_shape, partition_shape, np.dtype(np.int8), fill_value = 0.0)
     else:
         print(arg)
         raise NotImplementedError
